+++ getlino/setup_info.py
@@ -1,6 +1,6 @@
 SETUP_INFO = dict(
     name='getlino',
-    version='20.1',
+    version='20.1.1',
     install_requires=['click', 'virtualenv', 'jinja2'],
     # tests_require=['docker', 'atelier'],
     # test_suite='tests',