+++ docs/changes.rst
@@ -9,6 +9,11 @@ Changes in `getlino`
 Add the ciao projet to KNOWN_REPOS.
 Add 'sudo' to the certbot command.
 
+Fix some issues with creating user and database with the :cmd:`getlino startsite` .
+
+Released version 20.1 to PyPI
+Released version 20.1.1 to PyPI
+
 
 2019-11-23
 ==========